refactor: StringCalculator.kt 소스 파일에서 Operator Enum 클래스 분리
